# docs/swagger.yaml
basePath: /api/v1
definitions:
  main.CurrencyPrice:
    properties:
      ask:
        type: string
      bid:
        type: string
      high:
        type: string
      last:
        type: string
      low:
        type: string
      open:
        type: string
      timestamp:
        type: string
      volume:
        type: string
      volume_quote:
        type: string
    type: object
  main.HTTPBadRequestError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: 'Bad Request Error: the server cannot or will not process the request
          due to something that is perceived to be a client error (for example, malformed
          request syntax, invalid request message framing, or deceptive request routing).'
        type: string
    type: object
  main.HTTPFileNotFoundError:
    properties:
      code:
        example: 404
        type: integer
      message:
        example: 'File Not Found Error: Cannot retrieve the page that was requested.
          The following are some common causes of this error message: The requested
          file has been renamed. The requested file has been moved to another location
          and/or deleted.'
        type: string
    type: object
  main.HTTPInternalServerError:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: 'Internal Server Error: The server encountered an unexpected condition
          that prevented it from fulfilling the request. This error response is a
          generic '
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Crypto Server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Crypto Server API
  version: "1.0"
paths:
  /currency/{symbol}:
    get:
      consumes:
      - application/json
      description: Get the real-time crypto price for a currency symbol
      parameters:
      - description: Currency symbol (e.g. BTCUSDT, ETHBTC)
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: CurrencyPrice object
          schema:
            $ref: '#/definitions/main.CurrencyPrice'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.HTTPBadRequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.HTTPFileNotFoundError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.HTTPInternalServerError'
      summary: Get the real-time crypto price for a currency symbol
      tags:
      - Currency
  /currency/all:
    get:
      consumes:
      - application/json
      description: Get the real-time crypto prices for all supported currencies
      produces:
      - application/json
      responses:
        "200":
          description: Map of currency symbols and their real-time prices
          schema:
            additionalProperties:
              $ref: '#/definitions/main.CurrencyPrice'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.HTTPBadRequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.HTTPFileNotFoundError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.HTTPInternalServerError'
      summary: Get the real-time crypto prices for all supported currencies
      tags:
      - Currency
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"